# lib/api-zod/src/generated/types/generateIncomePlanBody.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
import type { GenerateIncomePlanBodyTime } from "./generateIncomePlanBodyTime";

export interface GenerateIncomePlanBody {
  /** Monthly income in USD */
  income: number;
  /** User's main skill */
  skill: string;
  /** Time available per day */
  time: GenerateIncomePlanBodyTime;
}