# lib/api-zod/src/generated/types/incomePlan.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
import type { Opportunity } from "./opportunity";

export interface IncomePlan {
  /** Monthly under-earning amount */
  under_earning_amount: string;
  /** Score 0-100 representing income optimization level */
  income_gap_score: string;
  /** The user's biggest income mistake */
  biggest_mistake: string;
  opportunities: Opportunity[];
  /** Day-by-day action plan for 7 days */
  seven_day_plan: string[];
  /** Potential monthly income increase */
  potential_increase: string;
  /** Yearly income being lost */
  yearly_loss: string;
}